Move LogInButton functionality to Navbar
Remove the LogInButton component and move its functionality to Navbar
